feature(viewattendant) Create ViewAttendant Component - Create ViewAttendant Component - Create function to make request to get attendants in attendantAction file - Add markup for viewing attendants. [Delivers #163655444]

+++ client/src/components/Products/Products.js
@@ -15,8 +15,11 @@ class Products extends Component {
     }
 
     deleteProduct(id) {
-      console.log(id)
-      this.props.deleteProduct(id);
+      if (confirm('Are you sure you want to delete this product?')){
+        console.log(id)
+        this.props.deleteProduct(id);
+      };
+      
     }
 
   render() {
